+++ app/src/main/java/com/example/mariaproj/Admin/ShowProduct.java
@@ -60,7 +60,7 @@ public class ShowProduct extends AppCompatActivity implements AdapterView.OnItem
                     volunteer.setPlace(c.getString(c.getColumnIndexOrThrow(COLUMN_VOLUNTEER_PLACE)));
                     volunteer.setPdescribtion(c.getString(c.getColumnIndexOrThrow(COLUMN_PLACE_DESCRIPTION)));
                     volunteer.setRequiredSup(c.getString(c.getColumnIndexOrThrow(COLUMN_REQUIRED_SUPPLIES)));
-                    volunteer.setRequiredNumOfVolunteers(c.getDouble(c.getColumnIndexOrThrow(COLUMN_NUM_OF_VOLUNTEERS)));
+                    volunteer.setRequiredNumOfVolunteers(c.getInt(c.getColumnIndexOrThrow(COLUMN_NUM_OF_VOLUNTEERS)));
                     volunteer.setNumOfRegisteredVolunteers(c.getInt(c.getColumnIndexOrThrow(COLUMN_REGISTERED_VOLUNTEERS)));
                     volunteer.setImageByte(c.getBlob(c.getColumnIndexOrThrow(COLUMN_PRODUCT_IMAGE)));
                     place_info[i] = new Volunteer(volunteer);
